requirements.txt
Django==2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.5
redis==2.10.6
